KAR-1935 Minor fixes to align with hardware

See merge request ska-telescope/mid-dish/ska-mid-dish-manager!316

+++ tests/acceptance/test_noise_diode_attributes.py
@@ -14,8 +14,8 @@ def reset_diode_params(dish_manager_proxy: tango.DeviceProxy):
     yield
     # restore defaults
     dish_manager_proxy.noiseDiodeMode = NoiseDiodeMode.OFF
-    dish_manager_proxy.periodicNoiseDiodePars = np.array([0, 0, 0], dtype=np.int64)
-    dish_manager_proxy.pseudoRandomNoiseDiodePars = np.array([0, 0, 0], dtype=np.int64)
+    dish_manager_proxy.periodicNoiseDiodePars = np.array([0, 0, 0], dtype=np.uint32)
+    dish_manager_proxy.pseudoRandomNoiseDiodePars = np.array([0, 0, 0], dtype=np.uint32)
 
 
 @pytest.mark.acceptance
@@ -23,8 +23,8 @@ def reset_diode_params(dish_manager_proxy: tango.DeviceProxy):
     ("attribute, valid_write_value"),
     [
         ("noiseDiodeMode", NoiseDiodeMode.PERIODIC),
-        ("periodicnoisediodepars", np.array([1, 2, 3], dtype=np.int64)),
-        ("pseudorandomnoisediodepars", np.array([1, 2, 3], dtype=np.int64)),
+        ("periodicnoisediodepars", np.array([1, 2, 3], dtype=np.uint32)),
+        ("pseudorandomnoisediodepars", np.array([1, 2, 3], dtype=np.uint32)),
     ],
 )
 def test_set_noise_diode_attribute(
